+++ src/main/java/com/divudi/ejb/PharmacyBean.java
@@ -1056,12 +1056,12 @@ public class PharmacyBean {
         return v;
     }
 
-    public double getOrderingQty(Item item, Department department) {
+    public float getOrderingQty(Item item, Department department) {
         double qty = 10;
         if (item instanceof Ampp) {
             qty /= item.getDblValue();
         }
-        return qty;
+        return (float)qty;
     }
 
     public double getMaximumPurchasePriceChange() {
@@ -1247,7 +1247,7 @@ public class PharmacyBean {
             vmpp.setName(vmp.getName() + " " + vmp.getCategory().getName() + " (" + issueUnitsPerPack + " " + packUnit.getName() + ")");
             vmpp.setPackUnit(packUnit);
             vmpp.setCreatedAt(Calendar.getInstance().getTime());
-            vmpp.setDblValue(issueUnitsPerPack);
+            vmpp.setDblValue((float)issueUnitsPerPack);
             getVmppFacade().create(vmpp);
         } else {
             vmpp.setRetired(false);
@@ -1274,7 +1274,7 @@ public class PharmacyBean {
             ampp = new Ampp();
             ampp.setAmp(amp);
             ampp.setName(amp.getName() + " " + issueUnitsPerPack + amp.getMeasurementUnit() + unit.getName());
-            ampp.setDblValue(issueUnitsPerPack);
+            ampp.setDblValue((float)issueUnitsPerPack);
             ampp.setMeasurementUnit(unit);
             ampp.setVmpp(vmpp);
             getAmppFacade().create(ampp);
@@ -1373,7 +1373,7 @@ public class PharmacyBean {
 //        m.put("t", BillType.PharmacyGrnBill);
 //        return getBillItemFacade().findDoubleByJpql(sql, m);
 //    }
-    public double getLastPurchaseRate(Item item, Department dept) {
+    public float getLastPurchaseRate(Item item, Department dept) {
         //System.out.println("getting last purchase rate");
         if (item instanceof Ampp) {
             item = ((Ampp) item).getAmp();
@@ -1394,9 +1394,9 @@ public class PharmacyBean {
         ItemBatch ii = getItemBatchFacade().findFirstBySQL(sql, m);
         // //System.err.println("d = " + ii.getPurcahseRate());
         if (ii != null) {
-            return ii.getPurcahseRate();
+            return (float)ii.getPurcahseRate();
         } else {
-            return 0.0;
+            return 0.0f;
         }
 
     }
@@ -1517,7 +1517,7 @@ public class PharmacyBean {
 
     }
 
-    public double getLastRetailRate(Item item, Department dept) {
+    public float getLastRetailRate(Item item, Department dept) {
         //System.out.println("getting last purchase rate");
         if (item instanceof Ampp) {
             item = ((Ampp) item).getAmp();
@@ -1536,9 +1536,9 @@ public class PharmacyBean {
         ItemBatch ii = getItemBatchFacade().findFirstBySQL(sql, m);
         // //System.err.println("d = " + ii.getPurcahseRate());
         if (ii != null) {
-            return ii.getRetailsaleRate();
+            return (float)ii.getRetailsaleRate();
         } else {
-            return 0.0;
+            return 0.0f;
         }
 
     }
